style(modeling) split square out from rectangle code
